chore: inital work to migrate to Query's alpha package setup
