+++ nx.json
@@ -1,16 +1,20 @@
 {
+  "$schema": "./node_modules/nx/schemas/nx-schema.json",
+  "affected": {
+    "defaultBase": "main"
+  },
   "tasksRunnerOptions": {
     "default": {
-      "runner": "@nrwl/nx-cloud",
+      "runner": "nx-cloud",
       "options": {
         "cacheableOperations": [
           "test:lib",
           "test:eslint",
           "test:types",
-          "build:types",
-          "build",
-          "rollup"
+          "test:build",
+          "build"
         ],
+        "parallel": 5,
         "accessToken": "OTI3Y2U3NGQtYzQ3ZC00ZmE3LWJjZWQtYTYxOTEyNmNiN2IyfHJlYWQtb25seQ=="
       }
     }
@@ -22,56 +26,45 @@
     }
   },
   "namedInputs": {
-    "globalBuildAffectingConfig": [
-      "{workspaceRoot}/babel.config.js",
-      "{workspaceRoot}/rollup.config.js",
-      "{workspaceRoot}/rollup.config.ts",
-      "{workspaceRoot}/tsconfig.json",
-      "{workspaceRoot}/tsconfig.base.json"
-    ],
-    "globalNonBuildAffectingConfig": [
-      "{workspaceRoot}/.eslintrc",
-      "{workspaceRoot}/jest-preset.js"
+    "sharedGlobals": [
+      "{workspaceRoot}/.browserslistrc",
+      "{workspaceRoot}/.eslintrc.cjs",
+      "{workspaceRoot}/babel.config.cjs",
+      "{workspaceRoot}/package.json",
+      "{workspaceRoot}/scripts/getRollupConfig.js",
+      "{workspaceRoot}/tsconfig.json"
     ],
     "default": [
+      "sharedGlobals",
       "{projectRoot}/**/*",
-      "globalBuildAffectingConfig",
-      "globalNonBuildAffectingConfig",
-      "!{projectRoot}/**/*.md",
-      "!{projectRoot}/**/build/**/*"
+      "!{projectRoot}/**/*.md"
     ],
     "public": [
       "default",
-      "!{workspaceRoot}/.eslintrc",
-      "!{workspaceRoot}/jest-preset.js",
-      "!{projectRoot}/**/?(*.)+(spec|test).[jt]s?(x)?(.snap)",
-      "!{projectRoot}/.eslintrc",
-      "!{projectRoot}/jest.config.js"
+      "{projectRoot}/build",
+      "{projectRoot}/dist",
+      "!{projectRoot}/.eslintrc.cjs",
+      "!{projectRoot}/tsconfig.eslint.json"
     ]
   },
   "targetDefaults": {
     "test:lib": {
-      "outputs": ["{projectRoot}/coverage"],
-      "inputs": ["default", "^public"]
+      "dependsOn": ["^build"],
+      "inputs": ["default", "^public"],
+      "outputs": ["{projectRoot}/coverage"]
     },
     "test:eslint": {
+      "dependsOn": ["^build"],
       "inputs": ["default", "^public"]
     },
     "test:types": {
-      "outputs": [
-        "{projectRoot}/build/**/*.d.ts",
-        "{projectRoot}/build/.tsbuildinfo"
-      ],
-      "inputs": ["default", "^public"],
-      "dependsOn": ["^test:types"]
+      "dependsOn": ["^build"],
+      "inputs": ["default", "^public"]
     },
-    "build:types": {
-      "outputs": [
-        "{projectRoot}/build/**/*.d.ts",
-        "{projectRoot}/build/.tsbuildinfo"
-      ],
+    "build": {
+      "dependsOn": ["^build"],
       "inputs": ["default", "^public"],
-      "dependsOn": ["^build:types"]
+      "outputs": ["{projectRoot}/build", "{projectRoot}/dist"]
     },
     "test:build": {
       "dependsOn": ["build"],
