+++ package.json
@@ -2,27 +2,27 @@
   "name": "form",
   "private": true,
   "repository": "https://github.com/tanstack/form.git",
+  "packageManager": "pnpm@8.5.1",
+  "type": "module",
   "scripts": {
-    "nx-reset": "nx reset",
     "clean": "pnpm --filter \"./packages/**\" run clean",
     "preinstall": "node -e \"if(process.env.CI == 'true') {console.log('Skipping preinstall...'); process.exit(1)}\" || npx -y only-allow pnpm",
-    "install:csb": "pnpm --filter \"./packages/**\" --prefer-offline install --no-frozen-lockfile",
+    "install:csb": "corepack enable && pnpm install --frozen-lockfile",
     "test": "pnpm run test:ci",
-    "test:ci": "nx affected --targets=test:lib,test:types,test:eslint,test:format --parallel=5",
-    "test:react:17": "nx affected --target=test:lib --parallel=5",
-    "test:eslint": "nx affected --target=test:eslint --parallel=5",
+    "test:ci": "nx run-many --exclude=examples/** --targets=test:format,test:eslint,test:lib,test:types,build,test:build",
+    "test:eslint": "nx affected --target=test:eslint",
     "test:format": "pnpm run prettier --check",
-    "test:lib": "nx affected --target=test:lib --parallel=5",
+    "test:lib": "nx affected --target=test:lib",
     "test:lib:dev": "pnpm --filter \"./packages/**\" run test:lib:dev",
-    "test:build": "nx run-many --target=test:build --projects=root",
-    "test:types": "nx affected --target=test:types --parallel=5",
-    "build": "nx run-many --target=build --projects=root",
-    "build:types": "nx affected --target=build:types --parallel=5",
-    "watch": "concurrently --kill-others \"rollup --config rollup.config.js -w\" \"pnpm run build:types --watch\"",
+    "test:build": "nx affected --target=test:build",
+    "test:types": "nx affected --target=test:types",
+    "build": "nx affected --target=build",
+    "build:all": "nx run-many --exclude=examples/** --target=build",
+    "watch": "pnpm run build:all && nx watch --all -- pnpm run build:all",
     "dev": "pnpm run watch",
-    "prettier": "prettier --plugin-search-dir . \"{packages,examples}/**/src/**/*.{md,js,jsx,ts,tsx,json,vue,svelte}\"",
+    "prettier": "prettier --plugin-search-dir . \"{packages,examples,scripts}/**/*.{md,js,jsx,cjs,ts,tsx,json,vue,svelte}\"",
     "prettier:write": "pnpm run prettier --write",
-    "cipublish": "ts-node scripts/publish.ts"
+    "cipublish": "node scripts/publish.js"
   },
   "nx": {
     "includedScripts": [
@@ -32,22 +32,19 @@
   },
   "namespace": "@tanstack",
   "devDependencies": {
-    "@babel/core": "^7.17.9",
-    "@babel/preset-env": "^7.16.11",
-    "@babel/preset-react": "^7.16.7",
-    "@babel/preset-typescript": "^7.16.7",
-    "@commitlint/parse": "^16.2.1",
-    "@faker-js/faker": "^6.3.1",
-    "@nrwl/nx-cloud": "latest",
-    "@rollup/plugin-babel": "^5.3.1",
-    "@rollup/plugin-commonjs": "22.0.1",
-    "@rollup/plugin-node-resolve": "^13.2.1",
-    "@rollup/plugin-replace": "^4.0.0",
-    "@testing-library/jest-dom": "^5.16.4",
-    "@testing-library/react": "^13.0.0",
-    "@testing-library/react-17": "npm:@testing-library/react@12.1.4",
-    "@testing-library/react-hooks": "^7.0.2",
-    "@testing-library/user-event": "14.4.3",
+    "@babel/core": "^7.21.8",
+    "@babel/preset-env": "^7.21.5",
+    "@babel/preset-react": "^7.18.6",
+    "@babel/preset-typescript": "^7.21.5",
+    "@commitlint/parse": "^17.6.5",
+    "@rollup/plugin-babel": "^6.0.3",
+    "@rollup/plugin-commonjs": "^25.0.0",
+    "@rollup/plugin-node-resolve": "^15.0.2",
+    "@rollup/plugin-replace": "^5.0.2",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^14.0.0",
+    "@testing-library/react-hooks": "^8.0.1",
+    "@testing-library/user-event": "^14.4.3",
     "@types/jest": "^26.0.4",
     "@types/luxon": "^2.3.1",
     "@types/node": "^17.0.25",
@@ -57,67 +54,56 @@
     "@types/testing-library__jest-dom": "^5.14.5",
     "@typescript-eslint/eslint-plugin": "^5.41.0",
     "@typescript-eslint/parser": "^5.41.0",
+    "@vitest/coverage-istanbul": "^0.27.1",
     "axios": "^0.26.1",
     "babel-eslint": "^10.1.0",
     "babel-jest": "^27.5.1",
-    "babel-plugin-transform-async-to-promises": "^0.8.18",
     "babel-preset-solid": "^1.5.4",
     "bundlewatch": "^0.3.2",
     "chalk": "^4.1.2",
     "concurrently": "^7.1.0",
+    "cpy-cli": "^4.2.0",
     "current-git-branch": "^1.1.0",
-    "eslint": "8.34.0",
-    "eslint-config-prettier": "^8.6.0",
-    "eslint-config-react-app": "^7.0.1",
-    "eslint-config-standard": "^17.0.0",
-    "eslint-config-standard-react": "^13.0.0",
-    "eslint-import-resolver-typescript": "^3.5.3",
-    "eslint-plugin-flowtype": "8.0.3",
+    "eslint": "^8.34.0",
+    "eslint-config-prettier": "^8.8.0",
+    "eslint-import-resolver-typescript": "^3.5.5",
+    "eslint-plugin-compat": "^4.1.4",
     "eslint-plugin-import": "^2.27.5",
-    "eslint-plugin-jsx-a11y": "^6.7.1",
-    "eslint-plugin-node": "^11.1.0",
-    "eslint-plugin-prettier": "^4.2.1",
-    "eslint-plugin-promise": "^6.1.1",
-    "eslint-plugin-react": "7.32.2",
+    "eslint-plugin-react": "^7.32.2",
     "eslint-plugin-react-hooks": "^4.6.0",
-    "eslint-plugin-standard": "^5.0.0",
     "git-log-parser": "^1.2.0",
-    "jest": "^27.5.1",
+    "jsdom": "^22.0.0",
     "jsonfile": "^6.1.0",
-    "luxon": "^2.3.2",
-    "nx": "15.9.2",
-    "nx-cloud": "^16.0.2",
-    "prettier": "^2.6.2",
-    "prettier-plugin-svelte": "^2.9.0",
+    "luxon": "^3.3.0",
+    "nx": "^16.3.1",
+    "nx-cloud": "^16.0.5",
+    "prettier": "^2.8.8",
+    "prettier-plugin-svelte": "^2.10.0",
+    "publint": "^0.1.12",
     "react": "^18.2.0",
     "react-17": "npm:react@^17.0.2",
     "react-dom": "^18.2.0",
     "react-dom-17": "npm:react-dom@^17.0.2",
-    "rimraf": "^3.0.2",
-    "rollup": "^2.70.2",
-    "rollup-plugin-preserve-directives": "0.1.0",
+    "rimraf": "^5.0.1",
+    "rollup": "^3.23.0",
+    "rollup-plugin-node-externals": "^6.1.0",
+    "rollup-plugin-preserve-directives": "^0.2.0",
     "rollup-plugin-size": "^0.2.2",
     "rollup-plugin-terser": "^7.0.2",
-    "rollup-plugin-visualizer": "^5.6.0",
+    "rollup-plugin-visualizer": "^5.9.0",
+    "rollup-preset-solid": "^2.0.1",
     "semver": "^7.3.8",
-    "solid-js": "^1.5.7",
-    "solid-testing-library": "^0.3.0",
+    "solid-js": "^1.6.13",
     "stream-to-array": "^2.3.0",
-    "ts-jest": "^27.1.1",
-    "ts-node": "^10.7.0",
+    "tsup": "^7.0.0",
+    "type-fest": "^3.11.0",
     "typescript": "^5.0.4",
-    "vue": "^3.2.33"
+    "vitest": "^0.27.1",
+    "vue": "^3.2.47"
   },
-  "bundlewatch": {
-    "files": [
-      {
-        "path": "packages/*/build/umd/*.production.js"
-      }
-    ]
-  },
-  "dependencies": {
-    "@tanstack/store": "^0.0.1-beta.84",
-    "fs-extra": "^11.1.1",
-    "rollup-plugin-dts": "^5.3.0"
+  "pnpm": {
+    "patchedDependencies": {
+      "@types/testing-library__jest-dom@5.14.5": "patches/@types__testing-library__jest-dom@5.14.5.patch"
+    }
   }
 }
