--- packages/form-core/.eslintrc
@@ -1,6 +1,0 @@
-{
-  "parserOptions": {
-    "project": "./tsconfig.json",
-    "sourceType": "module"
-  }
-}
