--- packages/form-core/jest.config.ts
@@ -1,4 +1,0 @@
-export default {
-  displayName: 'form-core',
-  preset: '../../jest-preset.js',
-}
