--- packages/react-form/jest.config.ts
@@ -1,5 +1,0 @@
-export default {
-  displayName: 'react-form',
-  preset: '../../jest-preset.js',
-  testMatch: ['<rootDir>/src/**/*.test.tsx', '<rootDir>/codemods/**/*.test.js'],
-}
