+++ packages/react-form/src/useField.tsx
@@ -1,3 +1,5 @@
+/* eslint-disable react/no-children-prop */
+
 import * as React from 'react'
 //
 import { useStore } from '@tanstack/react-store'
