--- project.json
@@ -1,24 +1,0 @@
-{
-  "name": "root",
-  "$schema": "../../node_modules/nx/schemas/project-schema.json",
-  "sourceRoot": "./",
-  "targets": {
-    "rollup": {
-      "command": "rollup --config rollup.config.js",
-      "outputs": [
-        "{workspaceRoot}/packages/*/build/**/*(.cjs|.mjs|.js)*",
-        "{workspaceRoot}/packages/*/build/stats*",
-        "{workspaceRoot}/packages/*/build/**/*.svelte"
-      ],
-      "inputs": ["default"]
-    },
-    "build": {
-      "command": "echo \"  @tanstack/form > All packages built! 📦\"",
-      "dependsOn": ["rollup", "^build:types", "^build"]
-    },
-    "test:build": {
-      "command": "bundlewatch"
-    }
-  },
-  "implicitDependencies": ["@tanstack/form-core", "@tanstack/react-form"]
-}
