+++ scripts/project.json
@@ -1,6 +1,6 @@
 {
+  "$schema": "../node_modules/nx/schemas/project-schema.json",
   "name": "scripts",
-  "$schema": "../../node_modules/nx/schemas/project-schema.json",
   "sourceRoot": "scripts",
   "targets": {
     "test:eslint": { "command": "eslint ./scripts" }
